feat: word frequency analyzer
