+++ include/Utils/Casting/Casting.hpp
@@ -1,0 +1,28 @@
+#ifndef CASTING_HPP
+#define CASTING_HPP
+
+#include "Dictionary/IDictionary.hpp"
+
+namespace Casting {
+    template <typename Key, typename Value, typename Impl>
+    const IDictionary<Key, Value>& toIDictionary(const Impl& impl);
+
+    /**
+     * @brief Converts a string identifier to a DictionaryType enum.
+     *
+     * This function attempts to find the given string `str` as a key in the
+     * global `stringDicitionaryTypeMap`. If successful, it returns the
+     * corresponding `DictionaryType` value.
+     *
+     * @param str The string representation of the dictionary type to convert.
+     * @return The corresponding `DictionaryType` enum value.
+     * @throw DictionaryTypeNotFound if the string does not correspond to a valid
+     * dictionary type in the map.
+     */
+    DictionaryType toDictionaryType(const std::string& str);
+}
+
+#include "Utils/Casting/Casting.impl.hpp"
+
+
+#endif
