+++ include/Utils/Casting/Casting.impl.hpp
@@ -1,0 +1,8 @@
+#include "Utils/Casting/Casting.hpp"
+
+namespace Casting {
+    template <typename Key, typename Value, typename Impl>
+    const IDictionary<Key, Value>& toIDictionary(const Impl& impl) {
+        return static_cast<const IDictionary<Key, Value>&>(impl);
+    }
+}
