+++ src/Utils/Casting/Casting.cpp
@@ -1,0 +1,11 @@
+#include "Utils/Casting/Casting.hpp"
+
+namespace Casting {
+    DictionaryType toDictionaryType(const std::string& str) {
+        try {
+            return stringDictionaryTypeMap.at(str);
+        } catch(const std::out_of_range& e) {
+            throw DictionaryTypeNotFound();
+        }
+    }
+}
